// vue3+ts项目/vue3-ts-cms/src/service/request/type.ts
import type { AxiosRequestConfig, AxiosResponse } from 'axios'

export interface IInterceptor {
  requestSuccessFn?: (config: AxiosRequestConfig) => AxiosRequestConfig
  requestFailureFn?: (err: any) => any
  responseSuccessFn?: <T = AxiosResponse>(config: T) => T
  responseFailureFn?: (err: any) => any
}

export interface IConfig extends AxiosRequestConfig {
  interceptor?: IInterceptor
}